boot-report.py: always rm /tmp/boot-report*

+++ boot-report.py
@@ -200,6 +200,8 @@ if maillog and mail_to:
     sys.stdout.flush()
     sys.stdout = stdout_save
     subprocess.check_output('cat %s | msmtp --read-envelope-from -t --' %maillog, shell=True)
+
+if maillog:
     if os.path.exists(maillog):
         if maillog.startswith('/tmp'):
             os.remove(maillog)
